Committing changes.

+++ matlab/printLatexPlot.m
@@ -11,7 +11,7 @@ function printLatexPlot(fileName, directory, width, options)
 %
 % COPYRIGHT : Neil D. Lawrence, 2010
   
-% SHEFFIELDML
+% GPMAT
 
   if nargin<4
     options = printLatexOptions;
@@ -42,55 +42,58 @@ function printLatexPlot(fileName, directory, width, options)
   else
     png = true;
   end
-  pos(3) = width/2.54;
-  pos(4) = height/2.54;
-  set(gcf, 'paperposition', pos);
   if isoctave
+    pos(3) = width/2.54;
+    pos(4) = height/2.54;
+    set(gcf, 'paperposition', pos);
     baseName = [pwd filesep fileName];
     newName = [pwd filesep directory filesep fileName];
     fprintf('Printing LaTeX plot ...\n');
-    print('-dtex', [fileName '.tex']);
-    if options.pdf
-      system(['epstopdf ' baseName '.eps --outfile ' baseName '.pdf']);
-      system(['mv ' baseName '.pdf ' newName '.pdf']);
-      if ~options.eps
-        system(['rm ' baseName '.eps']);
-      else
-        system(['mv ' baseName '.eps ' newName '.eps']);
-      end
-    end
-    
-    % Use another background file. 
-    if length(options.backgroundFile) == 0
-      backgroundFile = fileName;
-    else
+    if ~options.tikz
+      print('-dtex', [fileName '.tex']);
       if options.pdf
-        system(['rm ' baseName '.pdf'])
+        system(['epstopdf ' baseName '.eps --outfile ' baseName '.pdf']);
+        system(['mv ' baseName '.pdf ' newName '.pdf']);
+        if ~options.eps
+          system(['rm ' baseName '.eps']);
+        else
+          system(['mv ' baseName '.eps ' newName '.eps']);
+        end
       end
-      if options.eps
-        system(['rm ' baseName '.eps'])
-      end
-      backgroundFile = options.backgroundFile;
-    end
     
-    FID = fopen([baseName '.tex'], 'r');
-    fileStr = '';
-    lin = getline(FID);
-    i = 0;
-    while(lin>0)
-      i = i + 1;
-      lin = regexprep(lin, ['\\includegraphics{' fileName '}'], ...
-                      ['\includegraphics{' directory filesep backgroundFile '}']);
-      fileStr = [fileStr escapeText(lin) '\n'];
-      %disp(fileStr)
+      % Use another background file. 
+      if length(options.backgroundFile) == 0
+        backgroundFile = fileName;
+      else
+        if options.pdf
+          system(['rm ' baseName '.pdf'])
+        end
+        if options.eps
+          system(['rm ' baseName '.eps'])
+        end
+        backgroundFile = options.backgroundFile;
+      end
+    
+      FID = fopen([baseName '.tex'], 'r');
+      fileStr = '';
       lin = getline(FID);
+      i = 0;
+      while(lin>0)
+        i = i + 1;
+        lin = regexprep(lin, ['\\includegraphics{' fileName '}'], ...
+                        ['\includegraphics{' directory filesep backgroundFile '}']);
+        fileStr = [fileStr escapeText(lin) '\n'];
+        %disp(fileStr)
+        lin = getline(FID);
+      end
+      fclose(FID);
+      system(['rm -f ' fileName '.tex']);
+      FID = fopen([newName '.tex'], 'w');
+      fprintf(FID, fileStr);
+      fclose(FID);
+    else
+      print('-dtikz', [newName '.tex']);
     end
-    fclose(FID);
-    system(['rm -f ' fileName '.tex']);
-    FID = fopen([newName '.tex'], 'w');
-    fprintf(FID, fileStr);
-    fclose(FID);
-    
   else
     matfig2pgf('filename', [directory filesep backgroundFile '.tex'], ...
                'figwidth', width)
